requirements.txt
requests
google-adk
fastmcp
openai-agents